+++ lib/vcr/http_stubbing_adapters/fakeweb.rb
@@ -10,7 +10,8 @@ module VCR
       UNSUPPORTED_REQUEST_MATCH_ATTRIBUTES = [:body, :headers].freeze
       LOCALHOST_REGEX = %r|\Ahttps?://((\w+:)?\w+@)?(#{VCR::LOCALHOST_ALIASES.map { |a| Regexp.escape(a) }.join('|')})(:\d+)?/|i
 
-      VERSION_REQUIREMENT = '1.3.0'
+      MINIMUM_VERSION = '1.3.0'
+      MAXIMUM_VERSION = '1.3'
 
       def http_connections_allowed=(value)
         @http_connections_allowed = value