+++ spec/spec_helper.rb
@@ -36,5 +36,7 @@ RSpec.configure do |config|
 
   config.filter_run :focus => true
   config.run_all_when_everything_filtered = true
+
+  config.alias_it_should_behave_like_to :it_performs, 'it performs'
 end