+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Moved documentation from README to [Wiki](http://github.com/myronmarston/vcr/wiki).
 * Refactoring and code cleanup.
 * Fix InternetConnection.available? so that it memoizes correctly when a connection is not available.
+* Fix WebMock version checking to allow newly released 1.5.0 to be used without a warning.
 
 ## 1.2.0 (October 13, 2010)